Creating PersonalAccountService Methods

+++ Backend/Backend/Dtos/AccountDtos/WidthdrawDto.cs
@@ -5,6 +5,6 @@ namespace Backend.Dtos.AccountDtos
     public class WidthdrawDto
     {
         [Required]
-        public int Amount { get; set; }
+        public float Amount { get; set; }
     }
 }
